+++ pim-func-sim/libpimsim/src/pimCmd.h
@@ -249,7 +249,7 @@ public:
   {
     assert(cmdType == PimCmdEnum::REDSUM);
   }
-  pimCmdRedSum(PimCmdEnum cmdType, PimObjId src, int64_t* result, unsigned idxBegin, unsigned idxEnd)
+  pimCmdRedSum(PimCmdEnum cmdType, PimObjId src, int64_t* result, uint64_t idxBegin, uint64_t idxEnd)
     : pimCmd(cmdType), m_src(src), m_result(result), m_idxBegin(idxBegin), m_idxEnd(idxEnd)
   {
     assert(cmdType == PimCmdEnum::REDSUM_RANGE);
@@ -263,8 +263,8 @@ protected:
   PimObjId m_src;
   int64_t* m_result;
   std::vector<int> m_regionSum;
-  unsigned m_idxBegin = 0;
-  unsigned m_idxEnd = std::numeric_limits<unsigned>::max();
+  uint64_t m_idxBegin = 0;
+  uint64_t m_idxEnd = std::numeric_limits<uint64_t>::max();
 };
 
 //! @class  pimCmdBroadcast
